Applied non-critical post review fixes 2.0
