+++ ImageFeed/ServiceAccessLayer/OAuth2/OAuth2TokenResponseBody.swift
@@ -1,5 +1,5 @@
 //
-//  OAuthTokenResponseBody.swift
+//  OAuth2TokenResponseBody.swift
 //  ImageFeed
 //
 //  Created by Pavel Seleznev on 1/26/25.
@@ -7,7 +7,7 @@
 
 import Foundation
 
-struct OAuthTokenResponseBody: Decodable {
+struct OAuth2TokenResponseBody: Decodable {
     let accessToken: String
     
     private enum CodingKeys: String, CodingKey {
