+++ ImageFeed/ServiceAccessLayer/OAuth2/OAuth2TokenStorage.swift
@@ -7,10 +7,6 @@
 
 import Foundation
 
-protocol OAuth2TokenStorageProtocol {
-    var token: String? { get set }
-}
-
 // Provides data persistency for OAuth2Token using userDefaults storage
 final class OAuth2TokenStorage: OAuth2TokenStorageProtocol {
     private let userDefaults = UserDefaults.standard
