+++ ImageFeed/ServiceAccessLayer/OAuth2/OAuth2TokenStorageDelegate.swift
@@ -1,0 +1,12 @@
+//
+//  OAuth2TokenStorageProtocol.swift
+//  ImageFeed
+//
+//  Created by Pavel Seleznev on 1/26/25.
+//
+
+import Foundation
+
+protocol OAuth2TokenStorageProtocol {
+    var token: String? { get set }
+}
